print a char and create holberton.h
The function _write was create. This function puts a character in the standard output running equal to putchar. On success,
the number of bytes written is returned (zero indicates nothing was written). On error,
-1 is returned.
Additionally, the holberton.h file was created to store the prototypes and the definitions of the structures.