VERIFY THE NULL TERMINATION
in the last commit the null termination of a string was doing at the start
of the program and this do that program terminate with out print any character.
in this commit we evaluate the null termination only if this is after the % and
return (-1).
Also we move the evaluation of NULL format to the start of the function and eliminate
the validate.c function