ADD FUNCTION POR PRINT %d AND %i CASES
Add the call for the case where the function should prints %d and %i.
This call is evaluate in print_format.c with a array that compare the types.
Add the function por these cases printing all the option for an integer. The
functions were add in printers.c function.
In holberton.h add the prototypes for the new functions